requirements.txt
altgraph==0.17.3
bottle==0.12.25
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clr-loader==0.2.5
colorama==0.4.6
defusedxml==0.7.1
distlib==0.3.6
django-qrcode==0.3
docutils==0.20.1
filelock==3.12.2
fonttools==4.40.0
fpdf2==2.7.4
idna==3.4
pefile==2023.2.7
Pillow==9.5.0
platformdirs==3.5.3
proxy-tools==0.1.0
pycparser==2.21
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
pypiwin32==223
pypng==0.20220715.0
pythonnet==3.0.1
pywebview==4.2.2
pywin32==306
pywin32-ctypes==0.2.2
qrcode==7.4.2
requests==2.31.0
typing_extensions==4.6.3
urllib3==2.0.3
virtualenv==20.23.0
